Added weight area and helper class that does notwork yet.

+++ SDNDB/Weight/Tables/CardioLog.sql
@@ -2,7 +2,7 @@
     [ID]             INT             IDENTITY (1, 1) NOT NULL,
     [CardioItemID]   INT             NOT NULL,
     [Time]           DECIMAL (18, 2) NOT NULL,
-    [CaloriesBurned] INT             NOT NULL,
+    [CaloriesBurned] FLOAT (53)      NOT NULL,
     [WorkoutDate]    DATE            NOT NULL,
     [CreatedTime]    DATETIME        CONSTRAINT [DF_CardoLog_CreatedTime] DEFAULT (getdate()) NOT NULL,
     [Deleted]        BIT             CONSTRAINT [DF_CardoLog_Deleted] DEFAULT ((0)) NOT NULL,
@@ -12,4 +12,6 @@
     CONSTRAINT [FK_CardoLog_Items] FOREIGN KEY ([CardioItemID]) REFERENCES [Weight].[CardioItems] ([ID]),
     CONSTRAINT [FK_CardoLog_People] FOREIGN KEY ([CreatedBy]) REFERENCES [Gas].[People] ([ID])
 );
+
+
